Fixed namespaces. Unified ModeChange.
