+++ Assets/Scripts/Services/History/Actions.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9a7e2e0fe40545d47ad73714490d13a9
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
