+++ Assets/Scripts/Services/History/Actions/UserAction.cs
@@ -1,4 +1,4 @@
-﻿namespace Services.History {
+﻿namespace Services.History.Actions {
 	public interface UserAction {
 		void Execute();
 		void UnExecute();
