--- Assets/Scripts/Services/History/ModeAction.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 35b2ae4cab75ebf4fb91226de35bfa4e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
